--- packages/traceloop-sdk/tests/dataset/cassettes/test_columns_operations_vcr/test_create_dataset_with_columns.yaml
@@ -1,65 +1,0 @@
-interactions:
-- request:
-    body: '{"slug": "test-columns-1754926486", "name": "Test Columns Dataset", "description":
-      "Dataset for testing column operations", "columns": [{"slug": "name", "name":
-      "Name", "type": "string"}, {"slug": "price", "name": "Price", "type": "string"},
-      {"slug": "instock", "name": "InStock", "type": "string"}, {"slug": "rating",
-      "name": "Rating", "type": "string"}], "rows": [{"Name": "Product A", "Price":
-      "99.99", "InStock": "true", "Rating": "4.5"}, {"Name": "Product B", "Price":
-      "149.99", "InStock": "false", "Rating": "3.8"}, {"Name": "Product C", "Price":
-      "79.99", "InStock": "true", "Rating": "4.2"}]}'
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '599'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - python-requests/2.32.3
-      X-Traceloop-SDK-Version:
-      - 1.0.0
-    method: POST
-    uri: https://api-staging.traceloop.com/v2/datasets
-  response:
-    body:
-      string: '{"error":"row 0 has an invalid column slug: Rating"}'
-    headers:
-      CF-RAY:
-      - 96d8cc942fc1bca2-ATL
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '52'
-      Content-Type:
-      - application/json; charset=utf-8
-      Date:
-      - Mon, 11 Aug 2025 15:34:47 GMT
-      Permissions-Policy:
-      - geolocation=(self), microphone=()
-      Server:
-      - cloudflare
-      cf-cache-status:
-      - DYNAMIC
-      referrer-policy:
-      - strict-origin-when-cross-origin
-      strict-transport-security:
-      - max-age=7776000; includeSubDomains
-      via:
-      - kong/3.7.1
-      x-content-type:
-      - nosniff
-      x-kong-proxy-latency:
-      - '0'
-      x-kong-request-id:
-      - 70d1ed14504ab80b8ac008642897c762
-      x-kong-upstream-latency:
-      - '6'
-    status:
-      code: 400
-      message: Bad Request
-version: 1
