+++ packages/traceloop-sdk/tests/dataset/cassettes/test_dataset_operations/test_get_dataset_by_version.yaml
@@ -13,27 +13,33 @@ interactions:
       X-Traceloop-SDK-Version:
       - 1.0.0
     method: GET
-    uri: https://api-staging.traceloop.com/v2/datasets/product-inventory/versions/v1
+    uri: https://api-staging.traceloop.com/v2/datasets/nina-qa/versions/v1
   response:
     body:
-      string: '{"error":"Dataset not found"}'
+      string: 'New Column 1,New Column 2,New Column 3
+
+        hallo,0,false
+
+        '
     headers:
       CF-RAY:
-      - 96d8cca0db7fc987-IAD
+      - 96d9f09e0c493a78-FRA
       Connection:
       - keep-alive
       Content-Length:
-      - '29'
+      - '53'
       Content-Type:
-      - application/json; charset=utf-8
+      - text/csv
       Date:
-      - Mon, 11 Aug 2025 15:34:49 GMT
+      - Mon, 11 Aug 2025 18:54:09 GMT
       Permissions-Policy:
       - geolocation=(self), microphone=()
       Server:
       - cloudflare
       cf-cache-status:
       - DYNAMIC
+      content-disposition:
+      - attachment; filename="nina-qa-v1.csv"
       referrer-policy:
       - strict-origin-when-cross-origin
       strict-transport-security:
@@ -45,10 +51,10 @@ interactions:
       x-kong-proxy-latency:
       - '0'
       x-kong-request-id:
-      - 394257712bda8c4202545d6fbcd2d339
+      - b1758d02a9e56223e194eb1248ccf569
       x-kong-upstream-latency:
-      - '4'
+      - '50'
     status:
-      code: 404
-      message: Not Found
+      code: 200
+      message: OK
 version: 1
