+++ packages/traceloop-sdk/tests/dataset/cassettes/test_dataset_operations/test_publish_dataset.yaml
@@ -1,9 +1,9 @@
 interactions:
 - request:
-    body: '{"slug": "test-publish-dataset-1754926489", "name": "Test Publish Dataset",
+    body: '{"slug": "test-publish-dataset-1754938483", "name": "Test Publish Dataset",
       "description": "Dataset for testing publish functionality", "columns": [{"slug":
       "name", "name": "Name", "type": "string"}, {"slug": "value", "name": "Value",
-      "type": "string"}], "rows": [{"Name": "Test", "Value": "123"}]}'
+      "type": "string"}], "rows": [{"name": "Test", "value": "123"}]}'
     headers:
       Accept:
       - '*/*'
@@ -23,18 +23,19 @@ interactions:
     uri: https://api-staging.traceloop.com/v2/datasets
   response:
     body:
-      string: '{"error":"row 0 has an invalid column slug: Name"}'
+      string: '{"id":"cme7h291r003e01059xsmfmr8","slug":"test-publish-dataset-1754938483","name":"Test
+        Publish Dataset","description":"Dataset for testing publish functionality","columns":{"name":{"slug":"name","name":"Name","type":"string"},"value":{"slug":"value","name":"Value","type":"string"}},"created_at":"2025-08-11T18:54:44.559052077Z","updated_at":"2025-08-11T18:54:44.559052138Z","rows":[{"id":"cme7h291x003f0105xa3br809","row_index":1,"values":{"name":"Test","value":"123"},"created_at":"2025-08-11T18:54:44.56803782Z","updated_at":"2025-08-11T18:54:44.56803782Z"}]}'
     headers:
       CF-RAY:
-      - 96d8cca489ccbfd6-ATL
+      - 96d9f176ec30e7b5-FRA
       Connection:
       - keep-alive
       Content-Length:
-      - '50'
+      - '563'
       Content-Type:
       - application/json; charset=utf-8
       Date:
-      - Mon, 11 Aug 2025 15:34:50 GMT
+      - Mon, 11 Aug 2025 18:54:44 GMT
       Permissions-Policy:
       - geolocation=(self), microphone=()
       Server:
@@ -50,12 +51,12 @@ interactions:
       x-content-type:
       - nosniff
       x-kong-proxy-latency:
-      - '0'
+      - '1'
       x-kong-request-id:
-      - 52ed34501b964e1e1d2d3474b85b575f
+      - a6b3b4765aebf85d15d5f870ed6e211f
       x-kong-upstream-latency:
-      - '2'
+      - '17'
     status:
-      code: 400
-      message: Bad Request
+      code: 201
+      message: Created
 version: 1
