+++ packages/traceloop-sdk/tests/dataset/cassettes/test_rows_operations/test_create_dataset_and_add_rows.yaml
@@ -1,0 +1,64 @@
+interactions:
+- request:
+    body: '{"slug": "test-rows-1754938511", "name": "Test Rows Dataset", "description":
+      "Dataset for testing row operations", "columns": [{"slug": "name", "name": "Name",
+      "type": "string"}, {"slug": "age", "name": "Age", "type": "string"}, {"slug":
+      "active", "name": "Active", "type": "string"}], "rows": [{"name": "John", "age":
+      "25", "active": "true"}, {"name": "Jane", "age": "30", "active": "false"}, {"name":
+      "Bob", "age": "35", "active": "true"}]}'
+    headers:
+      Accept:
+      - '*/*'
+      Accept-Encoding:
+      - gzip, deflate
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '442'
+      Content-Type:
+      - application/json
+      User-Agent:
+      - python-requests/2.32.3
+      X-Traceloop-SDK-Version:
+      - 1.0.0
+    method: POST
+    uri: https://api-staging.traceloop.com/v2/datasets
+  response:
+    body:
+      string: '{"id":"cme7h2v4d003g0105grqt6cmn","slug":"test-rows-1754938511","name":"Test
+        Rows Dataset","description":"Dataset for testing row operations","columns":{"active":{"slug":"active","name":"Active","type":"string"},"age":{"slug":"age","name":"Age","type":"string"},"name":{"slug":"name","name":"Name","type":"string"}},"created_at":"2025-08-11T18:55:13.16549955Z","updated_at":"2025-08-11T18:55:13.16549961Z","rows":[{"id":"cme7h2v4j003h0105qpcmjmav","row_index":1,"values":{"active":"true","age":"25","name":"John"},"created_at":"2025-08-11T18:55:13.174259737Z","updated_at":"2025-08-11T18:55:13.174259737Z"},{"id":"cme7h2v4j003i0105yuu8qksb","row_index":2,"values":{"active":"false","age":"30","name":"Jane"},"created_at":"2025-08-11T18:55:13.174259737Z","updated_at":"2025-08-11T18:55:13.174259737Z"},{"id":"cme7h2v4j003j0105cfh7f91m","row_index":3,"values":{"active":"true","age":"35","name":"Bob"},"created_at":"2025-08-11T18:55:13.174259737Z","updated_at":"2025-08-11T18:55:13.174259737Z"}]}'
+    headers:
+      CF-RAY:
+      - 96d9f229ce2d68a5-FRA
+      Connection:
+      - keep-alive
+      Content-Length:
+      - '994'
+      Content-Type:
+      - application/json; charset=utf-8
+      Date:
+      - Mon, 11 Aug 2025 18:55:13 GMT
+      Permissions-Policy:
+      - geolocation=(self), microphone=()
+      Server:
+      - cloudflare
+      cf-cache-status:
+      - DYNAMIC
+      referrer-policy:
+      - strict-origin-when-cross-origin
+      strict-transport-security:
+      - max-age=7776000; includeSubDomains
+      via:
+      - kong/3.7.1
+      x-content-type:
+      - nosniff
+      x-kong-proxy-latency:
+      - '0'
+      x-kong-request-id:
+      - 700c8629469ac535bc8f3a5109b92034
+      x-kong-upstream-latency:
+      - '20'
+    status:
+      code: 201
+      message: Created
+version: 1
