--- packages/traceloop-sdk/tests/dataset/cassettes/test_rows_operations_vcr/test_create_dataset_and_add_rows.yaml
@@ -1,63 +1,0 @@
-interactions:
-- request:
-    body: '{"slug": "test-rows-1754926490", "name": "Test Rows Dataset", "description":
-      "Dataset for testing row operations", "columns": [{"slug": "name", "name": "Name",
-      "type": "string"}, {"slug": "age", "name": "Age", "type": "string"}, {"slug":
-      "active", "name": "Active", "type": "string"}], "rows": [{"Name": "John", "Age":
-      "25", "Active": "true"}, {"Name": "Jane", "Age": "30", "Active": "false"}, {"Name":
-      "Bob", "Age": "35", "Active": "true"}]}'
-    headers:
-      Accept:
-      - '*/*'
-      Accept-Encoding:
-      - gzip, deflate
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '442'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - python-requests/2.32.3
-      X-Traceloop-SDK-Version:
-      - 1.0.0
-    method: POST
-    uri: https://api-staging.traceloop.com/v2/datasets
-  response:
-    body:
-      string: '{"error":"row 0 has an invalid column slug: Name"}'
-    headers:
-      CF-RAY:
-      - 96d8cca909b498d4-ATL
-      Connection:
-      - keep-alive
-      Content-Length:
-      - '50'
-      Content-Type:
-      - application/json; charset=utf-8
-      Date:
-      - Mon, 11 Aug 2025 15:34:51 GMT
-      Permissions-Policy:
-      - geolocation=(self), microphone=()
-      Server:
-      - cloudflare
-      cf-cache-status:
-      - DYNAMIC
-      referrer-policy:
-      - strict-origin-when-cross-origin
-      strict-transport-security:
-      - max-age=7776000; includeSubDomains
-      via:
-      - kong/3.7.1
-      x-content-type:
-      - nosniff
-      x-kong-proxy-latency:
-      - '0'
-      x-kong-request-id:
-      - afaf4625d609e16ea7695f0938674b5b
-      x-kong-upstream-latency:
-      - '4'
-    status:
-      code: 400
-      message: Bad Request
-version: 1
