--- packages/traceloop-sdk/tests/dataset/test_rows_operations_vcr.py
@@ -1,91 +1,0 @@
-import pytest
-import tempfile
-import os
-import time
-
-
-@pytest.mark.vcr
-def test_create_dataset_and_add_rows(datasets):
-    """Test creating a dataset and adding rows using real API calls"""
-    try:
-        # Create a unique slug to avoid conflicts
-        unique_slug = f"test-rows-{int(time.time())}"
-        
-        # Create a simple CSV for the dataset
-        csv_content = """Name,Age,Active
-John,25,true
-Jane,30,false
-Bob,35,true"""
-        
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".csv", delete=False) as f:
-            f.write(csv_content)
-            csv_path = f.name
-        
-        try:
-            # Create dataset from CSV
-            dataset = datasets.from_csv(
-                file_path=csv_path,
-                slug=unique_slug,
-                name="Test Rows Dataset",
-                description="Dataset for testing row operations",
-            )
-            
-            assert dataset is not None
-            assert dataset.slug == unique_slug
-            assert len(dataset.columns) >= 3  # Name, Age, Active
-            assert len(dataset.rows) >= 0  # Allow any number of initial rows
-            
-        finally:
-            os.unlink(csv_path)
-            
-    except Exception as e:
-        # Allow for expected API errors during recording
-        assert ("Failed to create dataset" in str(e) or 
-                "401" in str(e) or "403" in str(e) or "409" in str(e))
-
-
-@pytest.mark.vcr  
-def test_dataset_row_operations_api_errors(datasets):
-    """Test handling of API errors for row operations"""
-    try:
-        # Try to get a non-existent dataset to record error response
-        dataset = datasets.get_by_slug("definitely-non-existent-dataset-12345")
-        # If we get here, the dataset unexpectedly exists, which is also valid for testing
-        assert dataset is not None
-        
-    except Exception as e:
-        # Should get a "dataset not found" type error
-        assert ("Failed to get dataset" in str(e) or "404" in str(e) or "401" in str(e))
-
-
-@pytest.mark.vcr
-def test_dataset_deletion(datasets):
-    """Test dataset deletion using real API calls"""
-    try:
-        # Create a dataset to delete
-        unique_slug = f"test-delete-{int(time.time())}"
-        
-        csv_content = """Name\nTestData"""
-        
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".csv", delete=False) as f:
-            f.write(csv_content)
-            csv_path = f.name
-        
-        try:
-            # Create dataset
-            dataset = datasets.from_csv(
-                file_path=csv_path,
-                slug=unique_slug, 
-                name="Test Delete Dataset",
-                description="Dataset for testing deletion",
-            )
-            
-            # Try to delete it
-            datasets.delete_by_slug(unique_slug)
-            
-        finally:
-            os.unlink(csv_path)
-            
-    except Exception as e:
-        # Allow for expected API errors during recording
-        assert ("Failed" in str(e) or "401" in str(e) or "403" in str(e) or "404" in str(e))
